Fixed missed cast to int for learning period. in tm detector too

+++ nab/detectors/numenta/numentaTM_detector.py
@@ -67,7 +67,7 @@ class NumentaTMDetector(NumentaDetector):
     self.model.enableInference({"predictedField": "value"})
 
     # Initialize the anomaly likelihood object
-    numentaLearningPeriod = math.floor(self.probationaryPeriod / 2.0)
+    numentaLearningPeriod = int(math.floor(self.probationaryPeriod / 2.0))
     self.anomalyLikelihood = anomaly_likelihood.AnomalyLikelihood(
       claLearningPeriod=numentaLearningPeriod,
       estimationSamples=self.probationaryPeriod-numentaLearningPeriod,
